(chore) Move Mosaic mock data to the root of mosaic container because they're used by Storybook too.
