+++ containers/mosaic/src/__tests__/components/Field/FormFieldDropdown/FormFieldDropdown.test.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { render, screen, cleanup, act, waitFor } from "@testing-library/react";
 
 import FormFieldDropdown from "@root/components/Field/FormFieldDropdown";
-import { optionsLibrary, getOptions } from "@root/mock";
+import { mockOptions, getOptions } from "@root/mock";
 
 const topFilms = [
 	{ label: "The Shawshank Redemption", value: "1994" },
@@ -81,15 +81,15 @@ describe("Dropdown component with options from DB", () => {
 							placeholder: "Placeholder test",
 						},
 					}}
-					value={optionsLibrary[7]}
+					value={mockOptions[7]}
 				/>,
 			);
 		});
-		await waitFor(async () => {
-			const input = getByRole<HTMLInputElement>("combobox");
-			expect(input).toBeInTheDocument();
-			expect(input).toHaveValue(optionsLibrary[7].label);
-		}, { timeout: 5000 });
+
+		await waitFor(() => {
+			const inputDropdown = getByRole("combobox") as HTMLInputElement;
+			expect(inputDropdown.value).toEqual(mockOptions[7].label);
+		}, { timeout: 3000 });
 	});
 
 	it("should render on the screen with options from DB", async () => {
@@ -117,7 +117,7 @@ describe("Dropdown component with options from DB", () => {
 
 		await waitFor(() => {
 			const singleSelectOptions = screen.getAllByRole("option");
-			expect(singleSelectOptions[0]).toHaveTextContent(optionsLibrary[0].label);
-		}, { timeout: 5000 });
+			expect(singleSelectOptions[0]).toHaveTextContent(mockOptions[0].label);
+		}, { timeout: 1000 });
 	});
 });
