+++ containers/mosaic/src/__tests__/mock/address.ts
@@ -1,5 +1,3 @@
-import type { AddressData } from "@root/components";
-
 /**
  * These suggestions are in the format that would be expected
  * from the Google Maps Places AutocompleteService.
@@ -163,54 +161,6 @@ export const mockGeocodedItems = [
 		},
 		place_id: "ChIJzeBE6AwFcUgR9MGNxwMF7Og",
 		types: ["premise"],
-	},
-];
-
-export const mockAddresses: AddressData = [
-	{
-		address1: "81 Sussex Gardens",
-		address2: "Hucclecote",
-		country: {
-			label: "United Kingdom",
-			value: "GB",
-		},
-		city: "Gloucester",
-		state: {
-			label: "Gloucestershire",
-			value: "GLS",
-		},
-		postalCode: "GL3 3SP",
-		types: [{ label: "Billing", value: "billing" }],
-	},
-	{
-		address1: "81 Sussex Gardens",
-		address2: "Hucclecote",
-		country: {
-			label: "United Kingdom",
-			value: "GB",
-		},
-		city: "Gloucester",
-		state: {
-			label: "Gloucestershire",
-			value: "GLS",
-		},
-		postalCode: "GL3 3SP",
-		types: [{ label: "Physical", value: "physical" }],
-	},
-	{
-		address1: "81 Sussex Gardens",
-		address2: "Hucclecote",
-		country: {
-			label: "United Kingdom",
-			value: "GB",
-		},
-		city: "Gloucester",
-		state: {
-			label: "Gloucestershire",
-			value: "GLS",
-		},
-		postalCode: "GL3 3SP",
-		types: [{ label: "Shipping", value: "shipping" }],
 	},
 ];
