--- containers/mosaic/src/__tests__/mock/options.ts
@@ -1,60 +1,0 @@
-import type { MosaicLabelValue } from "@root/types";
-
-export const mockOptions: MosaicLabelValue[] = [
-	{
-		label: "Dog",
-		value: "dog",
-	},
-	{
-		label: "Cat",
-		value: "cat",
-	},
-	{
-		label: "Horse",
-		value: "horse",
-	},
-	{
-		label: "Lion",
-		value: "lion",
-	},
-	{
-		label: "Tiger",
-		value: "tiger",
-	},
-	{
-		label: "Human",
-		value: "human",
-	},
-	{
-		label: "Cow",
-		value: "cow",
-	},
-	{
-		label: "Snake",
-		value: "snake",
-	},
-	{
-		label: "Bear",
-		value: "bear",
-	},
-	{
-		label: "Wolf",
-		value: "wolf",
-	},
-	{
-		label: "Koala",
-		value: "koala",
-	},
-	{
-		label: "Deer",
-		value: "deer",
-	},
-	{
-		label: "Giraffe",
-		value: "giraffe",
-	},
-	{
-		label: "Panda",
-		value: "panda",
-	},
-];
