+++ containers/sb-8/stories/components/Field/FormFieldCheckbox/FormFieldCheckbox.stories.tsx
@@ -4,8 +4,8 @@ import React, { useMemo } from "react";
 import type { FieldDef } from "@root/components/Field";
 
 import Form, { useForm } from "@root/components/Form";
-import { getOptions, optionsLibrary } from "@root/mock";
 import { commonFieldControls, renderButtons } from "../../../../utils";
+import { getOptions, mockOptions } from "@root/mock";
 
 export default {
 	title: "FormFields/FormFieldCheckbox",
@@ -35,7 +35,7 @@ export const Playground = ({
 					required,
 					disabled,
 					inputSettings: {
-						options: optionsType === "Synchronous" ? optionsLibrary : getOptions,
+						options: optionsType === "Synchronous" ? mockOptions : getOptions,
 					},
 					helperText,
 					instructionText,
@@ -92,7 +92,7 @@ const kitchenSinkFields: FieldDef[] = [
 		required: false,
 		disabled: false,
 		inputSettings: {
-			options: optionsLibrary,
+			options: mockOptions,
 		},
 		helperText: "Helper Text",
 		instructionText: "InstructionText",
@@ -104,7 +104,7 @@ const kitchenSinkFields: FieldDef[] = [
 		required: false,
 		disabled: true,
 		inputSettings: {
-			options: optionsLibrary,
+			options: mockOptions,
 		},
 		helperText: "Helper Text",
 		instructionText: "InstructionText",
